requirements.txt
# This is the requirements file

# DEVELOPMENT/PRODUCTION

python-dotenv
requests
pandas
pynytimes
sendgrid

# TEST
pytest

